Cleanup: Entfernte alte Firered- und Rogue-Dateien

+++ src/battle_controller_player.c
@@ -47,6 +47,7 @@
 #include "pokemon_summary_screen.h"
 #include "type_icons.h"
 #include "battle_main.h"
+#include "pokedex.h"
 
 static void PlayerBufferExecCompleted(u32 battler);
 static void PlayerHandleLoadMonSprite(u32 battler);
@@ -2340,7 +2341,7 @@ static void PlayerHandleOneReturnValue_Duplicate(u32 battler)
 static void PlayerHandleIntroTrainerBallThrow(u32 battler)
 {
     u32 backSpriteId = GetPlayerBackSpriteId();
-    const u32 *trainerPal = gTrainerBacksprites[backSpriteId].palette.data;
+    const u16 *trainerPal = gTrainerBacksprites[backSpriteId].palette.data;
     BtlController_HandleIntroTrainerBallThrow(battler, 0xD6F8, trainerPal, 31, Intro_TryShinyAnimShowHealthbox);
 }
